requirements.txt
Flask==3.1.1
numpy==2.2.6
tensorflow_intel==2.17.0
transformers==4.44.2
